Fixed unsafe type cast warnings in ApplicationCore on iOS
#preflight none
#rb trivial
#rnx

[CL 21704797 by Devin Doucette in ue5-main branch]

+++ Engine/Source/Runtime/ApplicationCore/Private/IOS/Accessibility/IOSAccessibilityCache.cpp
@@ -116,7 +116,7 @@ static void DumpAccessibilityStatsForwarder()
 
 					// Bounding rect
 					FBox2D Bounds = Widget->GetBounds();
-					const float Scale = [IOSAppDelegate GetDelegate].IOSView.contentScaleFactor;
+					const double Scale = [IOSAppDelegate GetDelegate].IOSView.contentScaleFactor;
 					Bounds.Min /= Scale;
 					Bounds.Max /= Scale;
